Use fetch request property wrapper
This replaces the original implementation in the AnnimalsNearYou view
with a @FetchRequest property wrapper. This also requires having access
to the Persistent store via the managedObjectContext environment
variable. This is set up in the ContentView and provided to the
AnimalsNearYou view. Note: the preview also needs access to the managed
object context, and in this case uses the preview version of the
Persistent store (which uses an in-memory store instead of the SQLite
store used by the shared store).

YouTrack Issue: ^PS-4