Use animal fetching service to fetch animals
This removes the dependency of the view model on the networking
implementation to fetch animals. Dependency inversion is used by
creating the AnimalsFetcher protocol and providing a service which
implements this protocol. The view model now only depends on the
protocol not the implementation details of how the data is fetched.

To provide data for previews, a mock implementation is also created
which implements the protocol. It doesn't fetch data from a network
service, just uses the data from the JSON file created for previews.

YouTrack Issue: ^PS-5