Ensure the view is updated on the main thread
The network service provided by the FetchAnimalsService retrieves data
on a background thread (since URLSession uses background threads when
making HTTP requests). Using @MainActor ensures code executed in this
class is run on the main thread - which is a requirement for all updates
to the User Interface (UI).

YouTrack Issue: ^PS-5