Allow infinite scrolling in view
Retrieving animals uses paging so only a small number are retrieved -
this avoids downloading significant numbers of animals from the network.
This originally only downloaded a single page of animals. Now when you
scroll to the end of the list, a new page of animals will be retrieved
from the network.

YouTrack Issue: ^PS-5