Use State and Binding to coordinate state value
This examples shows how a state value in the AnimalListView can be
passed down to the AnimalDetailsView. Also, the value used in the
AnimalDetailsView is bound to the state variable defined in the
AnimalListView. This menas the state value (isNavigationDisabled) is
shared between the views, and each view will know when it is modified.

YouTrack Issue: ^PS-8