Use StateObject and Observable/ObservedObject
This manages state using the ObservableObject class defined in
AnimalsNearYou. A StateObject is created in AnimalsListView and this is
passed down to the AnimalDetailsView which observes the object.

The functionality is similar to the previous State and Binding, however,
this uses a class object (instead of a type) to manage the state value
which is the Publised variable inside the NavigationState class.

Note also, since NavigationState is a class, it is passed down to
AnimalDetailsView without the need to prepend the '$' symbol - since
class initialisers return a reference.

YouTrack Issue: ^PS-8